+++ internal/app/crawler/proxy.go
@@ -11,19 +11,34 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package icache
+package crawler
 
 import (
-	"context"
-	"time"
+	"fmt"
+	"os"
 
-	"github.com/bsm/redislock"
+	_ "github.com/joho/godotenv/autoload"
 )
 
-type IRedis interface {
-	SetExpire(ctx context.Context, key string, expired time.Time) error
-	SAdd(ctx context.Context, key string, value string) error
-	SMembers(ctx context.Context, key string) ([]string, error)
-	Close() error
-	ObtainLock(ctx context.Context, key string, expire time.Duration) *redislock.Lock
+const (
+	WebScraping    = "WEB_SCRAPING"
+	WebScrapingUrl = "https://api.webscrapingapi.com/v1?api_key=%s"
+	ProxyCrawl     = "PROXY_CRAWL"
+	ProxyCrawlUrl  = "https://api.proxycrawl.com/?token=%s"
+)
+
+type proxy struct {
+	Type string
+}
+
+func (p *proxy) URI() string {
+	var url string
+	switch p.Type {
+	case ProxyCrawl:
+		url = ProxyCrawlUrl
+	case WebScraping:
+		url = WebScrapingUrl
+	}
+	token := os.Getenv(p.Type)
+	return fmt.Sprintf(url, token)
 }
